pyproject.toml
[tool.poetry]
name = "nudecrawler"
version = "0.1.0"
description = "Crawl telegra.ph searching for nudes!"
authors = ["Yaroslav Polyakov <yaroslaff@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/yaroslaff/nudecrawler"
repository = "https://github.com/yaroslaff/nudecrawler"
license = "MIT"
keywords = ["nsfw", "tits", "nudity-detection", "nsfw-recognition", "nudes", "onlyfans", "nude", "telegra-ph", "search", "crawler", "scraper", "spider", "scraping", "find", "web-scraping", "crawl", "scrape", "webscraping"]


[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.12.0"
transliterate = "^1.10.2"
pillow = "^9.4.0"
requests = "^2.28.2"
nudepy = "^0.5.1"
evalidate = "^1.0.2"
pytest = "^7.2.2"
nudenet = "^2.0.9"
python-daemon = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
